feat(CameraRig): add ability to move camera rig with touchpad
The touchpad can now be used to move the play area in the direction
the headset is facing. It can also be used to strafe left and right
in the current facing direction.

The new `SteamVR_TouchpadWalking` script is added to the `[CameraRig]`
prefab and adds a box collider and a rigidbody to the camera rig to
handle collisions.

A new example scene has been added to demonstrate this feature.